MOL-285: PR - Moved models to implementations folder and add type hints to models
