+++ tusd/hooks/implementations/hooks_utils.py
@@ -7,7 +7,7 @@ import psycopg2
 import psycopg2.extras
 
 from hooks.implementations.return_codes import DBERROR
-from hooks.models.DataFromDatabase import DataFromDatabase
+from hooks.implementations.models.DataFromDatabase import DataFromDatabase
 
 
 def read_tusd_event(step: str, input_data: TextIO, logger) -> dict:
